Aggiornamento per le funzionalita hash, controlli sulle date
