+++ .vscode/settings.json
@@ -4,5 +4,7 @@
         "stdio.h": "c",
         "prenotazione.h": "c",
         "stdlib.h": "c"
-    }
+    },
+    "makefile.configureOnOpen": false,
+    "C_Cpp.default.compilerPath": "C:/msys64/ucrt64/bin/gcc.exe"
 }
