--- Desktop/branchAnto/.vscode/settings.json
@@ -1,10 +1,0 @@
-{
-    "files.associations": {
-        "string.h": "c",
-        "stdio.h": "c",
-        "prenotazione.h": "c",
-        "stdlib.h": "c"
-    },
-    "makefile.configureOnOpen": false,
-    "C_Cpp.default.compilerPath": "C:/msys64/ucrt64/bin/gcc.exe"
-}
