--- Desktop/branchAnto/PSD/README.md
@@ -1,33 +1,0 @@
--- ADT Veicolo: Array di Struct (Veicolo) --
-	Tipo di veicolo: Stringa
-	Modello: Stringa
-	Colore: Stringa
-	Posizione: Stringa
-	*Orari disponibili: Float (?)
-		// Disponibilita : Bool (?)
-	Posti omologati: Intero
-	Accessibilita' per trasporto disabili: Char (Y per Si, N per No)
-	Tipo di combustibile: Stringa
-	Anno di immatricolazione: Intero
-	Costo noleggio: Float
-
--- Funzioni ADT Veicolo --
-	
-	stampaDettagli(veicolo) -> void  //Lo usiamo in modo iterativo per la stampa della disponibilitá dei veicoli (PUNTO 2)
-	modificaDisponibilita(veicolo, orarioScelto) -> void
-
----------------------------------------------------------------------------------------
-
--- Utente: stringa
-
----------------------------------------------------------------------------------------
-
--- ADT Prenotazione: --
-	Utente che prenotato: 
-
-
-
-
-
-
-
