--- Desktop/branchAnto/PSD/StoricoPrenotazioni.txt
@@ -1,3 +1,0 @@
-
-Anto-20/05/2025-8.00-9.30-16727-Fiat 500-AB123CD-0-0
-Anto-20/05/2025-13.30-15.00-28807-Mercedes Class V-WX901GH-9-3
