--- Desktop/branchAnto/PSD/utente.txt
@@ -1,2 +1,0 @@
-Fra
-Anto
