--- Desktop/branchAnto/Veicoli.txt
@@ -1,10 +1,0 @@
-Auto;Fiat 500;Rosso;AB123CD;4;Benzina;2021;35;
-SUV;Ford Kuga;Nero;XY987ZT;5;Ibrido;2022;60;
-Moto;Ducati Monster;Blu;GH654IJ;2;Benzina;2020;40;
-Furgone;Opel Vivaro;Bianco;MN321OP;3;Diesel;2023;75;
-Berlina;Audi A4;Grigio;QR456UV;5;Diesel;2021;90;
-Utilitaria;Peugeot 208;Verde;ST789WX;5;Benzina;2022;45;
-Elettrica;Tesla Model Y;Bianco;YZ012AB;5;Elettrico;2023;110;
-Furgoncino;Volkswagen Caddy;Blu;CD345EF;5;Diesel;2022;55;
-Sportiva;Ferrari Portofino;Rosso;UV678KL;2;Benzina;2023;250;
-Minivan;Mercedes Class V;Nero;WX901GH;7;Diesel;2021;80;
