+++ Orari.txt
@@ -1,1 +1,1 @@
-08.00;09.30;0;09.45;11.15;0;11.30;13.00;0;13.30;15.00;0;15.15;16.45;0;17.00;18.30;0;18.45;20.00;0;20.15;22.00;0
+08.00;09.30;0;09.45;11.15;0;11.30;13.00;0;13.30;15.00;0;15.15;16.45;0;17.00;18.30;0;18.45;20.00;0;20.15;22.00;0;
