+++ Veicolo.c
@@ -1,3 +1,8 @@
+/*ATTENZIONE: Nel codice sono presenti comandi system
+              Implementati per la pulizia del terminale
+              Potrebbero creare conflitti o problemi di vari natura
+              "Annullarli" nel caso della presenza di quest'ultimi*/
+
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
@@ -5,7 +10,7 @@
 
 /*-- FUNZIONI RELATIVE AI VEICOLI --*/
 
-/*
+/*---------------------------------------------------------------------------------------------------------------- 
  * Funzione: riempiVeicoli
  * -----------------------
  * Prende in input l'oggetto veicolo e il puntatore al file .txt contenente i dati dei veicoli
@@ -33,29 +38,35 @@
  * Effetti collaterali:
  *      Modifica il contenuto nell'oggetto veicolo
  *      Se il file è vuoto, la funzione potrebbe dare comportamento indefinito senza opportuni controlli
+ * ---------------------------------------------------------------------------------------------------------------- 
  */
 
- void riempiVeicoli (veicolo v, char* nomefile){
+ void riempiVeicoli (veicolo v){
 
     FILE *file;
     char buffer [200];
+    static int i=0;
+    int j = 0;
 
-    file = fopen ("nomefile", "r");
+    file = fopen ("Veicoli.txt", "r");
 
     if (file == NULL){
-
-        printf ("Errore nell'apertura del file.");
+        system("cls | clear");
+        perror ("Errore nell'apertura del file.");
         exit (1);
     }
-
-    while (fgets (buffer, sizeof (buffer), file) != '\n'){
+    do{
+        fgets (buffer, sizeof (buffer), file);
+        j++;
+    }while (j <= i);
 
         char* token = strtok (buffer, ";");
+        controlloToken (token, v, file);
 
         v->tipoVeicolo = malloc (strlen (token)+1 * sizeof (char));
 
         if (v->tipoVeicolo == NULL){
-
+            system("cls | clear");
             perror ("ERRORE TIPO VEICOLO");
             exit (1);
         }
@@ -63,11 +74,12 @@
         strcpy (v->tipoVeicolo, token);
 
         token = strtok (NULL, ";");
+        controlloToken (token, v, file);
 
         v->modello = malloc (strlen (token)+1 * sizeof (char));
 
         if (v->modello == NULL){
-
+            system("cls | clear");
             perror ("ERRORE MODELLO");
             free (v->tipoVeicolo);
             exit(1);
@@ -76,11 +88,12 @@
         strcpy (v->modello, token);
 
         token = strtok (NULL, ";");
+        controlloToken (token, v, file);
 
         v->colore = malloc (strlen (token)+1 * sizeof (char));
 
         if (v->colore == NULL){
-
+            system("cls | clear");
             perror ("ERRORE COLORE");
             free (v->modello);
             free (v->tipoVeicolo);
@@ -90,11 +103,12 @@
         strcpy (v->colore, token);
 
         token = strtok (NULL, ";");
+        controlloToken (token, v, file);
 
         v->targa = malloc (strlen (token)+1 * sizeof (char));
 
         if (v->targa == NULL){
-
+            system("cls | clear");
             perror ("ERRORE TARGA");
             free (v->modello);
             free (v->tipoVeicolo);
@@ -105,15 +119,17 @@
         strcpy (v->targa, token);
 
         token = strtok (NULL, ";");
+        controlloToken (token, v, file);
 
         v->postiOmologati = atoi (token);
 
         token = strtok (NULL, ";");
+        controlloToken (token, v, file);
 
         v->Combustibile = malloc (strlen(token)+1 * sizeof (char));
 
         if (v->Combustibile == NULL){
-
+            system("cls | clear");
             perror ("ERRORE COMBUSTIBILE");
             free (v->modello);
             free (v->tipoVeicolo);
@@ -122,25 +138,31 @@
             exit (1);
         }
 
+        strcpy (v->Combustibile, token);
+
         token = strtok (NULL, ";");
+        controlloToken (token, v, file);
 
         v->annoDiImmatricolazione = atoi (token);
 
         token = strtok (NULL, ";");
+        controlloToken (token, v, file);
 
         v->CostoNoleggioOrario = atoi (token);
-    }
+    
 
-    fclose (file);
+        riempiOrari (v);
 
-    if (file != EOF){
-
+    
+    if (fclose (file) != 0){
+        system("cls | clear");
         perror ("Errore nella chiusura del file.");
         exit (1);
     }
+    i++;
 }
 
-/*
+/*---------------------------------------------------------------------------------------------------------------- 
  * Funzione: stampaVeicoli
  * -----------------------
  * Stampa a video i dati di un oggetto veicolo tra cui anche gli intervalli orari con la loro disponibilità (richiamo ad altre due funzioni) 
@@ -166,17 +188,15 @@
  * 
  * Effetti collaterali:
  *      Nessun effetto collaterale
+ * ---------------------------------------------------------------------------------------------------------------- 
  */
 
  void stampaVeicolo (veicolo v){
-
-    printf("VEICOLO:\nTipo Veicolo: %s\nModello: %s\nColore: %s\nTarga: %s\nPosti Omologati: %d\nCombustibile: %s\nAnno di immatricolazione: %d\nCosto Noleggio: %.2f€/h", v->tipoVeicolo, v->modello, v->colore, v->targa, v->postiOmologati, v->Combustibile,v->annoDiImmatricolazione, v->CostoNoleggioOrario);
-
-    printf ("Fasce Orarie: ");
-    stampaOrari (v);
+    printf("Tipo Veicolo: %s\nModello: %s\nColore: %s\nTarga: %s\nPosti Omologati: %d\nCombustibile: %s\nAnno di immatricolazione: %d\nCosto Noleggio: %.2f euro/h\n\n", v->tipoVeicolo, v->modello, v->colore, v->targa, v->postiOmologati, v->Combustibile,v->annoDiImmatricolazione, v->CostoNoleggioOrario);
+    
 }
 
-/*
+/*---------------------------------------------------------------------------------------------------------------- 
  * Funzione: liberaVeicoli
  * -----------------------
  *  Libera la memoria dell'oggetto veicolo
@@ -201,27 +221,21 @@
  * 
  * Effetti collaterali:
  *      L'oggetto veicolo non ha più dati presenti in memoria
+ * ---------------------------------------------------------------------------------------------------------------- 
  */
 
 void liberaVeicoli (veicolo v){
 
-    for (int i=0; i<10; i++){
-
-        free (v->annoDiImmatricolazione);
         free (v->colore);
         free (v->Combustibile);
         free (v->modello);
         free (v->targa);
         free (v->tipoVeicolo);
-        free (v->orari);
-    }
-
-    free (v);
 }
 
 /*-- FUNZIONI RELATIVE AL NOLEGGIO --*/
 
-/*
+/*---------------------------------------------------------------------------------------------------------------- 
  * Funzione: costoNoleggio
  * -----------------------
  * Calcola quanto costa il noleggio in un orario scelto dall'utente con eventuale sconto 
@@ -248,6 +262,7 @@ void liberaVeicoli (veicolo v){
  * 
  * Effetti collaterali: 
  *      Nessun effetto collaterale
+ * ---------------------------------------------------------------------------------------------------------------- 
  */
 
 float costoNoleggio (veicolo v, int k){
@@ -271,7 +286,7 @@ float costoNoleggio (veicolo v, int k){
     return (tempoNoleggio * v->CostoNoleggioOrario) * verificaSconto(v, k);
 }
 
-/*
+/*---------------------------------------------------------------------------------------------------------------- 
  * Funzione: verificaSconto
  * -----------------------
  * La funzione verifica la possibilità di un sconto in determinati intervalli orari prestabiliti e restituisce un tipo float
@@ -299,6 +314,7 @@ float costoNoleggio (veicolo v, int k){
  *      
  * Effetti collaterali:
  *      Nessun effetto collaterale 
+ * ---------------------------------------------------------------------------------------------------------------- 
  */
 
 float verificaSconto (veicolo v, int k){ // Restituisce float tra 0.0 a 1.0 che corrisponde alla percentuale di sconto (1- verificaSconto);
@@ -313,7 +329,7 @@ float verificaSconto (veicolo v, int k){ // Restituisce float tra 0.0 a 1.0 che 
 
 /*-- FUNZIONI RELATIVE AGLI ORARI --*/
 
-/*
+/*---------------------------------------------------------------------------------------------------------------- 
  * Funzione: riempiOrari
  * -----------------------
  * Prende in input l'oggetto veicolo e il puntatore al file .txt contenente i dati degli intervalli orari
@@ -341,51 +357,57 @@ float verificaSconto (veicolo v, int k){ // Restituisce float tra 0.0 a 1.0 che 
  * Effetti collaterali:
  *      Modifica il contenuto nell'oggetto veicolo
  *      Se il file è vuoto, la funzione potrebbe dare comportamento indefinito senza opportuni controlli
+ * ---------------------------------------------------------------------------------------------------------------- 
  */
 
-void riempiOrari (veicolo v, char* nomefile){
+void riempiOrari (veicolo v){
 
     FILE *file;
     char buffer [200];
 
-    file = fopen ("nomefile", "r");
+    file = fopen ("Orari.txt", "r");
 
     if (file == NULL){
-
+        system("cls | clear");
         perror ("Errore nell'apertura del file.");
         exit (1);
     }
 
-    while (fgets (buffer, sizeof (buffer), file) != '\n'){
+        
+
+        while (fgets (buffer, sizeof (buffer), file) != NULL){
 
         char* token = strtok (buffer, ";");
+        controlloToken (token, v, file);
 
-            for (int k=0; k<8; k++){
+                for (int k=0; k<8; k++){
 
                 v->orari[k].inizio = atof (token);
 
                 token = strtok (NULL, ";");
+                controlloToken (token, v, file);
 
                 v->orari[k].fine = atof (token);
 
                 token = strtok (NULL, ";");
+                controlloToken (token, v, file);
 
-                v->orari[k].Disponibilità = atoi (token);
-            }
-    }
+                v->orari[k].Disponibilita = atoi (token);
 
-        rewind (file);
+                token = strtok (NULL, ";");
+                controlloToken (token, v, file);
+                }
+        }
 
-        file = fclose (file);
-
-        if (file != EOF){
-
+        if (fclose (file) != 0){
+            system("cls | clear");
             perror ("Errore nella chiusura del file.");
             exit (1);
         }
     }
 
-/*
+
+/*---------------------------------------------------------------------------------------------------------------- 
  * Funzione: verificaDisponibilità
  * -----------------------
  * Verifica che un veicolo in un determinato orario sia disponibile attraverso un modulo 2 (se 1 allora non disponibile altrimenti disponibile)
@@ -411,12 +433,13 @@ void riempiOrari (veicolo v, char* nomefile){
  *      Nessun valore
  * 
  * Effetti collaterali:
- *      Nessun effetto collaterale       
+ *      Nessun effetto collaterale   
+ * ----------------------------------------------------------------------------------------------------------------     
  */
 
-void verificaDisponibilità (veicolo v, int k){
+void stampaDisponibilita (veicolo v, int k){
 
-    if ((v->orari[k].Disponibilità)%2 == 1){
+    if ((v->orari[k].Disponibilita)%2 == 1){
 
         printf ("Non Disponibile\n");
     }
@@ -426,7 +449,13 @@ void verificaDisponibilità (veicolo v, int k){
     }
 }
 
-/*
+int verificaDisponibilita(veicolo v, int indiceOrario){
+    if (v->orari[indiceOrario].Disponibilita == 0){
+        return 1;
+    }
+    return 0;
+}
+/*---------------------------------------------------------------------------------------------------------------- 
  * Funzione: modificaDisponibilità
  * -----------------------
  * Va a modificare nel campo disponibilità il suo valore a 1 quando chiamata
@@ -452,16 +481,17 @@ void verificaDisponibilità (veicolo v, int k){
  * 
  * Effetti collaterali:
  *       Cambiato il valore nel campo Disponibilità della struct Orario
+ * ---------------------------------------------------------------------------------------------------------------- 
  */
 
-void modificaDisponibilità (veicolo v, int k){
+void modificaDisponibilita (veicolo v, int k){
 
-    v->orari[k].Disponibilità = 1;
+    v->orari[k].Disponibilita = 1;
 }
 
-/*
+/*-----------------------------------------------------------------------------------------------------------------
  * Funzione: stampaOrari
- * -----------------------
+ * ----------------------------------------------------------------------------------------------------------------
  * Stampa a video l'intervallo orario scelto dall'utente con la sua disponibilità (Non disponibile o Disponibile)
  * 
  * Specifica sintattica:
@@ -484,15 +514,60 @@ void modificaDisponibilità (veicolo v, int k){
  *      Nessun valore
  * 
  * Effetti collaterali:
- *      Nessun effetto collaterale      
+ *      Nessun effetto collaterale     
+ * ---------------------------------------------------------------------------------------------------------------- 
  */
 
 void stampaOrari (veicolo v){
 
         for (int k=0; k<8; k++){
 
-        printf ("%.2f-%.2f ", v->orari[k].inizio, v->orari[k].fine);
+        printf (" (%d) %.2f-%.2f ", k, v->orari[k].inizio, v->orari[k].fine);
 
-        verificaDisponibilità(v, k);
+        stampaDisponibilita(v, k);
         }
+}
+
+/*--Funzioni di supporto--*/
+
+/*-----------------------------------------------------------------------------------------------------------------
+ * Funzione: controllotoken
+ * ----------------------------------------------------------------------------------------------------------------
+ * Controllo di eventuali errori nella funzione strtok della libreria string.h 
+ * 
+ * Specifica sintattica:
+ *      void controllotoken(char*, veicolo, FILE*)-> void
+ *
+ * Parametri:
+ *      token: stringa
+ *      v: oggetto veicolo
+ *      file: puntatore a file
+ * 
+ * Specifica semantica:
+ *      controllotoken(token, v, file)-> void 
+ *       
+ * Pre-condizione:
+ *      token deve esistere
+ *       
+ * Post-condizione:
+ *      Se token == NULL, libera memoria e chiude file
+ * 
+ * Ritorna: 
+ *      Nessun valore     
+ * 
+ * Effetti collaterali:
+ *      Nessun effetto collaterale
+ * ---------------------------------------------------------------------------------------------------------------- 
+ */
+
+void controlloToken (char* token, veicolo v, FILE* file){
+
+    if (token == NULL){
+        system("cls | clear");
+        printf("Errore nella lettura dei dati.\n");
+        liberaVeicoli(v); // Nuova funzione per deallocare
+        fclose(file);
+        exit(1);
+    }
+
 }
